requirements.txt
# Core Flask
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5

# Real-time communication
Flask-SocketIO==5.3.6
eventlet==0.33.3

# AI Assistant dependencies - Opzionali (per IA locale)
SpeechRecognition==3.10.0
pyttsx3==2.90
# pyaudio==0.2.11  # Commentato - può causare problemi su alcuni sistemi

# Local AI & ML dependencies - Opzionali
# transformers==4.35.2  # Commentato - molto pesante
# torch==2.1.1          # Commentato - molto pesante  
# sentence-transformers==2.2.2  # Commentato - molto pesante
numpy==1.24.3
regex==2023.10.3

# PocketSphinx for offline speech recognition - Opzionale
# pocketsphinx==5.0.0   # Commentato - opzionale

# Security & Rate Limiting
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
cryptography==41.0.7

# Database
SQLAlchemy==2.0.23

# Data processing & Export
XlsxWriter==3.1.9
reportlab==4.0.6
python-dateutil==2.8.2

# System monitoring
psutil==5.9.6

# HTTP requests
requests==2.31.0

# Development dependencies
python-dotenv==1.0.0

# Optional: Production server
gunicorn==21.2.0

# Optional: Redis for production rate limiting
redis==5.0.1
